models moved to domain module, parser to adapetr/parser
