+++ src/adapter/parser/mod.rs
@@ -1,0 +1,3 @@
+mod parser;
+
+pub use parser::*;
