+++ src/data/mod.rs
@@ -1,7 +1,5 @@
-mod table;
 mod table_builder;
 mod column_builder;
 
-pub use table::*;
 pub use table_builder::*;
 pub use column_builder::*;
