+++ src/domain/metric.rs
