+++ src/domain/operator.rs
