+++ src/domain/query.rs
