+++ src/domain/symbol_metric.rs
