--- src/parser/mod.rs
@@ -1,17 +1,0 @@
-mod parser;
-mod query;
-mod expr;
-mod operator;
-mod metric;
-mod symbol_metric;
-mod time_unit;
-mod time_spec;
-
-pub use parser::*;
-pub use query::*;
-pub use expr::*;
-pub use operator::*;
-pub use metric::*;
-pub use symbol_metric::*;
-pub use time_unit::*;
-pub use time_spec::*;
